Move perform_request under storage since it is bound to storage requests.
