+++ src/azure/storage/mod.rs
@@ -2,3 +2,4 @@ pub mod client;
 pub mod container;
 pub mod blob;
 pub mod table;
+mod rest_client;
